+++ packages/plugin-js-packages/src/lib/runner/runner.ts
@@ -1,17 +1,28 @@
 import path from 'node:path';
-import type { RunnerFunction } from '@code-pushup/models';
+import type {
+  RunnerFunction,
+  TableColumnObject,
+  TableRowObject,
+} from '@code-pushup/models';
 import {
   asyncSequential,
+  capitalize,
   executeProcess,
+  formatAsciiTable,
+  logger,
   objectFromEntries,
+  objectToEntries,
+  pluralizeToken,
 } from '@code-pushup/utils';
 import {
   type AuditSeverity,
   type DependencyGroup,
   type FinalJSPackagesPluginConfig,
+  type PackageAuditLevel,
   type PackageJsonPath,
   type PackageManagerId,
   dependencyGroups,
+  packageAuditLevels,
 } from '../config.js';
 import { dependencyGroupToLong } from '../constants.js';
 import { packageManagers } from '../package-managers/package-managers.js';
@@ -54,6 +65,8 @@ async function processOutdated(
   depGroups: DependencyGroup[],
   packageJsonPath: PackageJsonPath,
 ) {
+  logger.info('Looking for outdated packages ...');
+
   const pm = packageManagers[id];
   const { stdout } = await executeProcess({
     command: pm.command,
@@ -62,9 +75,13 @@ async function processOutdated(
     ignoreExitCode: true, // outdated returns exit code 1 when outdated dependencies are found
   });
 
+  const normalizedResult = pm.outdated.unifyResult(stdout);
+  logger.info(
+    `Detected ${pluralizeToken('outdated package', normalizedResult.length)} in total`,
+  );
+
   const depTotals = await getTotalDependencies(packageJsonPath);
 
-  const normalizedResult = pm.outdated.unifyResult(stdout);
   return depGroups.map(depGroup =>
     outdatedResultToAuditOutput(
       normalizedResult,
@@ -88,6 +105,10 @@ async function processAudit(
     supportedAuditDepGroups.includes(group),
   );
 
+  logger.info(
+    `Auditing packages for ${pluralizeToken('dependency group', compatibleAuditDepGroups.length)} (${compatibleAuditDepGroups.join(', ')}) ...`,
+  );
+
   const auditResults = await asyncSequential(
     compatibleAuditDepGroups,
     async (depGroup): Promise<[DependencyGroup, AuditResult]> => {
@@ -104,6 +125,8 @@ async function processAudit(
   const resultsMap = objectFromEntries(auditResults);
   const uniqueResults = pm.audit.postProcessResult?.(resultsMap) ?? resultsMap;
 
+  logAuditSummary(uniqueResults);
+
   return compatibleAuditDepGroups.map(depGroup =>
     auditResultToAuditOutput(
       // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
@@ -113,4 +136,67 @@ async function processAudit(
       auditLevelMapping,
     ),
   );
+}
+
+function logAuditSummary(
+  results: Partial<Record<DependencyGroup, AuditResult>>,
+): void {
+  const { totalCount, countsPerLevel } = aggregateAuditResults(results);
+  const formattedLevels = objectToEntries(countsPerLevel)
+    .filter(([, count]) => count > 0)
+    .map(([level, count]) => `${count} ${level}`)
+    .join(', ');
+
+  logger.info(
+    [
+      `Found ${pluralizeToken('vulnerability', totalCount)} in total`,
+      formattedLevels && `(${formattedLevels})`,
+    ]
+      .filter(Boolean)
+      .join(' '),
+  );
+
+  if (!logger.isVerbose()) {
+    return;
+  }
+  logger.debug(
+    formatAsciiTable({
+      columns: [
+        { key: 'depGroup', label: 'Dep. group' },
+        ...[...packageAuditLevels, 'total'].map(
+          (level): TableColumnObject => ({
+            key: level,
+            label: capitalize(level),
+            align: 'right',
+          }),
+        ),
+      ],
+      rows: objectToEntries(results).map(
+        ([depGroup, result]): TableRowObject => ({
+          depGroup,
+          ...result?.summary,
+        }),
+      ),
+    }),
+  );
+}
+
+function aggregateAuditResults(
+  results: Partial<Record<DependencyGroup, AuditResult>>,
+) {
+  const totalCount = Object.values(results).reduce(
+    (acc, { vulnerabilities }) => acc + vulnerabilities.length,
+    0,
+  );
+  const countsPerLevel = Object.values(results).reduce<
+    Record<PackageAuditLevel, number>
+  >(
+    (acc, { summary }) =>
+      objectFromEntries(
+        packageAuditLevels.map(level => [level, acc[level] + summary[level]]),
+      ),
+    objectFromEntries(packageAuditLevels.map(level => [level, 0])),
+  );
+
+  return { totalCount, countsPerLevel };
 }